docsrs: enable feature gate
And do some refactor.
