+++ Cargo.toml
@@ -25,3 +25,7 @@ tonic = { version = "0.6.2", optional = true }
 sha2 = "0.9.0"
 rand = "0.8.3"
 tokio = { version = "1", features = ["macros", "rt", "io-util"] }
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
